make it possible to set max number of lines
The maximum number of label lines that appear below a file's icon is
still hardcoded to 2.  However, it can be now changed at compile time
by setting the value of the NLINES enum value.

For example, setting NLINES to 4 makes up to 4 lines of labels of file
names to be drawn on the window.  This also makes the icons to be placed
further apart.