add _CONTROL_GOTO; handle key presses in xfilesctl
The `_CONTROL_GOTO` property is watched by XFiles; and, whenever it
changes, XFiles tries to change its current working directory to the
path provided by this property.  The user can then set this property
with xprop(1), and change the current directory from dmenu(1), for
example (using dmenu as some kind of URL bar).

That feature was copied from Suckless' surf(1), which uses a `_SURF_GO`
property instead.  I could name that property's atom `_XFILES_GO`
instead, for the sake of regularity; but using a different atom for each
program can overload the database of interned Atoms.  So I propose for
using this `_CONTROL_GOTO` atom instead for future applications.

There is no standard or common practice on the name of the atom used to
set the path/URI/name of current working directory or current working
document.  (Especially because that few programs use such technique;
the only programs I know doing that are surf and xfiles itself).

The `_CONTROL_CWD` (CWD stands for "Current Working Directory/Document")
is the equivalent to surf's `_SURF_URI` property, set by the program to
the name of the current open document or directory.

Another new feature was to handle keypresses by the XFiles control
script (xfilesctl) rather than by xfiles itself.  The script can then be
modified by the user to do whatever the user wants on the selected
files.  Examples provided on the default script are handling the
"F5/Ctrl-R", "Ctrl-L" and "Delete" key presses.

That change required removing the code handling the WIDGET_PARENT
action; for it is now handled by xfilesctl.