unobfuscate type definition
In widget.h, replace
	typedef struct Widget *Widget;
with
	typedef struct Widget Widget;

Defining a name as a pointer to a structure hides the fact that it is a
pointer (and that it can be compared to NULL, or that it is subject to
pointer arithmetic, for example).  To make obvious the pointer nature of
the type, while still hiding its implementation as an opaque entity, it
is a better style to not typedef pointers.