improve snippet readability
Although it is clear from the code that `dpy` is a `Display *`, that may
not be obvious in a code snippet without its surrounding context.

It may be the influence of LISP style on me, but I'm slowly abandoning
UNIX overly abbreviated naming into full (but short) names.  It should
not be hard to write four extra characters after all (most text editors
feature completion anyway).