widget: fix mouse selection with {Caps,Num}Lock activated
previously, mouse selection was activated when the motion state was
exactly either Button1, Button1 + Shift, or Button1 + Control.  But
if the user had Capslock on, and were pressing Shift, the state was
Button1 + Shift + CapsLock, which did not triggered the selection.

Now only the existence of Button1 is checked in the XMotionEvent(3)
state field.

Also, rename the variable "state" to "event".