operate on highlighted file if Shift is pressed
This commit partially reverts ade9c55 (2023-12-04), which made
^KEY commands operate only on selected files.  Previously they
operated on selected files or, if no file was selected, on the
highlighted file instead.

This commit restores that previous behavior, but only when the
Shift key is also pressed.

This commit also deprecates the ^S-KEY commands used for Shift
modified key-presses.  Now command names are not prefixed with
"S-" if Shift is pressed.

GLOSSARY:
• selected files: files manually selected with the left-button
  mouse click or the Space key.  They get blue, in the default
  color scheme.
• highlighted file: single file used to anchor keyboard motion
  (with arrow keys) and is opened when the user presses Enter.
  It gets underlined in white, in the default color scheme.