widget event loop refactored
nextevent() was rewritten to ignore filtered events and get new event
until an unfiltered one is found, which is returned to the event loop
routine.  So event loops don't need to call nextevent() and then call
filter_event() (previously processevent()) and then continue the loop
if event is filtered.

nextevent() now returns the type of the unfiltered event, allowing us
to get an event and select its type on a single statement.

Two additional types of events have been defined to make it easier to
select events: CloseNotify (the window was closed, by user or from an
error) and TimeoutNotify (timed out).

This:

	for (;;) {
		if (nextevent(widget, &ev, timeout)) {
			/* timed out */
			...
			continue;
		}
		switch (processevent(widget, &ev)) {
		case WIDGET_CLOSE:
		case WIDGET_ERROR:
			/* application closed */
			return;
		case WIDGET_INTERNAL:
			/* filtered event */
			continue;
		}
		switch (ev.type) {
		case ButtonPress:
			...
			break;
		case PropertyNotify:
			...
			break;
		}
	}

Is now this:

	for (;;) switch (nextevent(widget, &ev, timeout)) {
	case CloseNotify:
		/* application closed */
		return;
	case TimeoutNotify:
		/* timed out */
		...
		continue;
	case ButtonPress:
		...
		continue;
	case PropertyNotify:
		...
		continue;
	}

Much simpler; and it saved us one level of indentation!

I prefered the "continue" rather than "break" in the switch statement
to make it clear we are not breaking the loop at the same line of the
switch statement, but continuing it.

Also, I standardized boolean constants.  libX11's Bool/True/False are
used favor of stdbool.h's bool/true/false constants.  Both were being
used indistinctly throughout ./widget.c.

Also, a delay on scrolling in selmode() is fixed.