add struct ctrldnd_data; change ctrldnd_drop() interface
`struct ctrldnd_data` holds the target (aka format), size in bytes, and
raw data in char[] for a dnd selection.

`ctrldnd_drop()` can drop a selection in several formats.  Previously it
was given three arrays: formats, converted data, and sizes.  And the
programmer required to keep their sizes and members in sync.  That is
a bad interface.

Also improve manuels.