ctrldnd: simplify atoms interning
An atom's value is kept by X the server, not the client.  Even if we
call some ctrldnd_*() function a second time with a different display
value, the atoms will have the same value as before.  There is no need
to intern atoms at each call.  Then, I made them static, file-global, and
interned only once.

I also ignored errors from XInternAtoms().  If it fails, X server is
low on memory or had some error.  The program crashing is the lesser
of the troubles in such situation.