remove unused enum; remove unneeded abstractions
I rewrote two statements literally, rather than abstracting them out and
referring to them by name.  No meaningful name describes their result;
the literal statements are clearer than whatever name could have been
assigned to them; the statements are too small and only used once; etc.

Having a name is a privilege; and some things are not worth having a
presence in the namespace.  Nor are they worth to populate a whole layer
of abstraction, and to become a cognitive load for the programmer when
reading the code and jumping between functions.